requirements.txt
treamlit==1.12.2
matplotlib==3.5.1
matplotlib-base==3.5.1
matplotlib-inline==0.1.2
numpy==1.21.5
pandas==1.4.2
pickleshare==0.7.5